Fix minor deleteCache() bugs and add some checks

+++ upload/admin/model/catalog/attribute.php
@@ -398,7 +398,10 @@ class ModelCatalogAttribute extends Model {
 	}
 
 	public function deleteCache($attribute_id, $store_id = null) : void {
-			
+		
+		$products 	= [];
+		$categories = [];
+		
 		if ($store_id === null) {
 			$store_id = $this->session->data['store_id'];
 		}
@@ -421,22 +424,27 @@ class ModelCatalogAttribute extends Model {
 				FROM " . DB_PREFIX . "product_to_category
 				WHERE product_id IN(" . implode(',',array_column($products, 'product_id')). ")
 					AND store_id = '" . (int) $store_id . "'
-			");
+			")->rows;
 		}
 
 		foreach ($languages as $language) {
 			$language_id = $language['language_id'];
-			// Pproduct cache
-			foreach ($products as $product) {
-				$productCacheName 	= "product.store_{$store_id}.language_{$language_id}." . (floor($product['product_id'] / 100)) . "00.product_" . $product['product_id'];
-				$this->cache->delete($productCacheName);
+			// Product cache
+			if (!empty($products)) {
+				foreach ($products as $product) {
+					$productCacheName 	= "product.store_{$store_id}.language_{$language_id}." . (floor($product['product_id'] / 100)) . "00.product_" . $product['product_id'];
+					$this->cache->delete($productCacheName);
+				}
 			}
-			foreach ($categories as $category) {
-				$category_id = (int) $category['category_id'];
 
-				// Filter cache
-				$filterCacheName = "category.store_{$store_id}.language_{$language_id}." . (floor($category_id / 100)) . "00.filters_{$category_id}";
-				$this->cache->delete($filterCacheName);
+			if (!empty($categories)) {
+				foreach ($categories as $category) {
+					$category_id = (int) $category['category_id'];
+	
+					// Filter cache
+					$filterCacheName = "category.store_{$store_id}.language_{$language_id}." . (floor($category_id / 100)) . "00.filters_{$category_id}";
+					$this->cache->delete($filterCacheName);
+				}
 			}
 		}
 	}
