+++ upload/admin/model/catalog/filter.php
@@ -623,6 +623,9 @@ class ModelCatalogFilter extends Model {
 
 	public function deleteCache($filter_id, $store_id = null) : void {
 
+		$products 	= [];
+		$categories = [];
+
 		if ($store_id === null) {
 			$store_id = (int) $this->session->data['store_id'];
 		}
@@ -636,6 +639,7 @@ class ModelCatalogFilter extends Model {
 			WHERE filter_id = '" . $filter_id . "'
 				AND store_id = '" . $store_id . "'
 		")->rows;
+
 		$products = $this->db->query("
 			SELECT
 				DISTINCT product_id
@@ -647,15 +651,19 @@ class ModelCatalogFilter extends Model {
 		foreach ($languages as $language) {
 			$language_id = $language['language_id'];
 			// Delete filters cache
-			foreach ($categories as $category) {
-				$filterCacheName = "category.store_{$store_id}.language_{$language_id}." . (floor($category['category_id'] / 100)) . "00.filters_{$category['category_id']}";
-				$this->cache->delete($filterCacheName);
+			if (!empty($products)) {
+				foreach ($categories as $category) {
+					$filterCacheName = "category.store_{$store_id}.language_{$language_id}." . (floor($category['category_id'] / 100)) . "00.filters_{$category['category_id']}";
+					$this->cache->delete($filterCacheName);
+				}
 			}
-
-			foreach ($products as $product) {
-				// Delete product cache
-				$productCacheName 	= "product.store_{$store_id}.language_{$language_id}." . (floor($product['product_id'] / 100)) . "00.product_{$product['product_id']}";
-				$this->cache->delete($productCacheName);
+			
+			if (!empty($categories)) {
+				foreach ($products as $product) {
+					// Delete product cache
+					$productCacheName 	= "product.store_{$store_id}.language_{$language_id}." . (floor($product['product_id'] / 100)) . "00.product_{$product['product_id']}";
+					$this->cache->delete($productCacheName);
+				}
 			}
 		}
 	}
